refactor: package 구조 수정 #42
