+++ src/main/java/com/parkez/reservation/domain/enums/LockStrategy.java
@@ -1,5 +1,5 @@
 package com.parkez.reservation.domain.enums;
 
 public enum LockStrategy {
-    DEFAULT, OPTIMISTIC, PESSIMISTIC;
+    DEFAULT;
 }
