--- src/main/java/com/parkez/reservation/domain/repository/pessimistic/PessimisticLockReservationRepository.java
@@ -1,7 +1,0 @@
-package com.parkez.reservation.domain.repository.pessimistic;
-
-import com.parkez.reservation.domain.entity.Reservation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PessimisticLockReservationRepository extends JpaRepository<Reservation, Long> {
-}
