--- src/main/java/com/parkez/reservation/service/distributed/DistributedLockReservationService.java
@@ -1,4 +1,0 @@
-package com.parkez.reservation.service.distributed;
-
-public class DistributedLockReservationService {
-}
